requirements.txt
# Core dependencies
click>=8.0.0              # CLI framework
requests>=2.28.0          # HTTP client for Ollama API
pyyaml>=6.0              # Configuration file parsing
rich>=13.0.0             # Terminal formatting and progress bars
gitpython>=3.1.0         # Git integration
watchdog>=3.0.0          # File monitoring
pygments>=2.14.0         # Syntax highlighting

# Optional dependencies for enhanced features
python-magic>=0.4.27     # File type detection (optional)
chardet>=5.0.0           # Character encoding detection (optional)

# Development dependencies (install with pip install -e .[dev])
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
pre-commit>=3.0.0
flake8>=6.0.0

# Documentation dependencies (install with pip install -e .[docs])
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
sphinx-click>=4.0.0